refactor: :recycle: move form-edit-register component to app folder
