+++ src/app/user/shelters/_components/index.ts
@@ -1,0 +1,2 @@
+export * from "./shelters-list-table";
+export * from "./form-edit-register";
